% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataDLAG.R
\name{createDataDLAG}
\alias{createDataDLAG}
\title{Simulate a data set with exposures measured over time such that there is an effect
of exposure 1 and 2 as well as an interaction between them}
\usage{
createDataDLAG(
  n = 200,
  p = 5,
  t = 37,
  q = 5,
  corr_exposures = 0.5,
  corr_time = 0.95,
  betaMain = NULL,
  betaInt = NULL
)
}
\arguments{
\item{n}{Sample size of the model}

\item{p}{Number of exposures}

\item{q}{Number of additional covariates}

\item{corr_exposures}{Correlation between exposures at a point in time}

\item{corr_time}{Correlation within exposures across time}

\item{betaInt}{Optional list of true distributed lag surfaces for each interaction
If left NULL, exposures 1 and 2 will have an interaction effect.}

\item{beta}{Optional list of true distributed lag curves for each exposure.
If left NULL, exposures 1 and 2 will have an effect on the outcome.}
}
\description{
Simulate a data set with exposures measured over time such that there is an effect
of exposure 1 and 2 as well as an interaction between them
}
