% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLAGbayesAlpha.R
\name{DLAGbayesAlpha}
\alias{DLAGbayesAlpha}
\title{Estimate effects of repeated exposures using distributed lag models and controls average posterior
inclusion probabilities for null associations}
\usage{
DLAGbayesAlpha(
  y,
  x,
  c = NULL,
  nScans = 2000,
  nBurn = 1000,
  thin = 2,
  PCAthresholdBasis = 0.95,
  PCAthresholdMain = 1,
  PCAthresholdInt = 0.999,
  VariableSel = TRUE,
  Basis = "PCA",
  alphaMain = 0.05,
  alphaInt = 0.01
)
}
\arguments{
\item{y}{The outcome to be analyzed}

\item{x}{A list where each element is a matrix containing the values
of each exposure over time}

\item{c}{An n by q matrix of additional, time invariant covariates to adjust for}

\item{nScans}{The number of MCMC scans to run}

\item{nBurn}{The number of MCMC scans that will be dropped as a burn-in}

\item{thin}{This number represents how many iterations between each scan
that is kept}

\item{PCAthresholdBasis}{This number represents what percentage of variation we want our basis functions
from PCA to capture from the original exposures}

\item{PCAthresholdMain}{The proportion of variation to keep in the main effect DLAG curves. We recommend
leaving this parameter fixed at 1}

\item{PCAthresholdInt}{The proportion of variation to keep in the interaction DLAG surfaces We recommend
leaving this parameter fixed at 0.999}

\item{VariableSel}{Whether to perform variable selection on main effect distributed lag surfaces}

\item{Basis}{Indicates what type of basis functions to use. The default is PCA, which then
uses the PCA threshold variable to determine how many basis functions to keep.
The user may also input "Cubic" to get Cubic polynomial basis functions}

\item{alphaMain}{The level at which we aim to restrict the posterior inclusion probabilities
for main effects with no association with the outcome}

\item{alphaInt}{The level at which we aim to restrict the posterior inclusion probabilities
for interactions with no association with the outcome}
}
\value{
A list of values that contain the posterior inclusion probabilities for the main effect
        distributed lags and interaction distributed lag surfaces. The list also contains the
        posterior distribution for the distributed lag parameters as well as the basis functions
        used to estimate their distributed lag function. These posterior distributions do not
        represent the distributed lag surfaces themselves, but these can be obtained using the
        other functions in the package. See PosteriorMainDLAG() and PosteriorInteractionDLAG().
}
\description{
Estimate effects of repeated exposures using distributed lag models and controls average posterior
inclusion probabilities for null associations
}
