% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{DLAGbayesAlphaBoth}
\alias{DLAGbayesAlphaBoth}
\title{Estimate effects of repeated exposures using distributed lag models}
\usage{
DLAGbayesAlphaBoth(
  nScans = 2000,
  nBurn = 1000,
  thin = 2,
  PCAthresholdBasis = 0.95,
  y,
  x,
  c = NULL,
  PCAthresholdMain = 1,
  PCAthresholdInt = 0.999,
  VariableSel = TRUE,
  Basis = "TPS",
  alphaMain = 0.05,
  alphaInt = 0.01
)
}
\arguments{
\item{nScans}{The number of MCMC scans to run}

\item{nBurn}{The number of MCMC scans that will be dropped as a burn-in}

\item{thin}{This number represents how many iterations between each scan
that is kept}

\item{y}{The outcome to be analyzed}

\item{x}{A list where each element is a matrix containing the values
of each exposure over time}

\item{c}{An n by q matrix of additional, time invariant covariates to adjust for}

\item{VariableSel}{Whether to perform variable selection on main effect distributed lag surfaces}

\item{Basis}{Indicates what type of basis functions to use. The default is PCA, which then
uses the PCA threshold variable to determine how many basis functions to keep.
The user may also input "Cubic" to get Cubic polynomial basis functions}

\item{PCAthreshold}{This number represents what percentage of variation we want our basis functions
from PCA to capture from the original exposures}
}
\value{
A list of values that contain the posterior inclusion probabilities for the main effect
        distributed lags and interaction distributed lag surfaces. The list also contains the
        posterior distribution for the distributed lag parameters as well as the basis functions
        used to estimate their distributed lag function. These posterior distributions do not
        represent the distributed lag surfaces themselves, but these can be obtained using the
        other functions in the package.
}
\description{
Estimate effects of repeated exposures using distributed lag models
}
\examples{

There will be an example here
}
