% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PosteriorInteractionDLAG}
\alias{PosteriorInteractionDLAG}
\title{Extract the full posterior distribution for an interaction distributed lag}
\usage{
PosteriorInteractionDLAG(DLAGmodel, whichExposures)
}
\arguments{
\item{DLAGmodel}{A fitted model from the DLAGbayes function}

\item{whichExposures}{a vector indicating which exposures the posterior
distribution of the interaction should be calculated.
Each must be between 1 and the number of exposures.}
}
\value{
An array containing the posterior distribution of the distributed lag. The dimensions of
        the array is NC x NI x T x T, where NC is the number of chains run, NI is the number of iterations
        saved per chain, and T is the number of time points we observe each exposure at.
}
\description{
Extract the full posterior distribution for an interaction distributed lag
}
\examples{

There will be an example here
}
