% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PosteriorMainDLAG}
\alias{PosteriorMainDLAG}
\title{Extract the full posterior distribution for a main effect distributed lag}
\usage{
PosteriorMainDLAG(DLAGmodel, whichExposure)
}
\arguments{
\item{DLAGmodel}{A fitted model from the DLAGbayes function}

\item{whichExposure}{an integer indicating which exposure the posterior
distribution of the main effect lag should be calculated.
Must be between 1 and the number of exposures}
}
\value{
An array containing the posterior distribution of the distributed lag. The dimensions of
        the array is NC x NI x T, where NC is the number of chains run, NI is the number of iterations
        saved per chain, and T is the number of time points we observe the exposure at.
}
\description{
Extract the full posterior distribution for a main effect distributed lag
}
\examples{

There will be an example here
}
